backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
supabase>=1.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pandas>=2.0.0
aiohttp>=3.8.0
requests>=2.28.0
python-docx>=1.0.0